Codewars.Tests/6kyu/TitleCaseTests.cs
using NUnit.Framework;

namespace Codewars.Tests
{
    [TestFixture]
    public class TitleCaseTests
    {
        [TestCase("a clash of KINGS", "a an the of", "A Clash of Kings")]
        [TestCase("THE WIND IN THE WILLOWS", "The In", "The Wind in the Willows")]
        public void MyTest(string sampleTitle, string sampleMinorWords, string expected)
        {
            Assert.AreEqual(expected, TitleCase.ToTitleCase(sampleTitle, sampleMinorWords));
        }
        [Test]
        public void MyTest2()
        {
            Assert.AreEqual("", TitleCase.ToTitleCase(""));
        }
        [Test]
        public void MyTest3()
        {
            Assert.AreEqual("The Quick Brown Fox", TitleCase.ToTitleCase("the quick brown fox"));
        }
    }
}